+++ dev/backend/haomibo/src/main/java/com/nuctech/ecuritycheckitem/repositories/SysUserGroupLookupRepository.java
@@ -17,8 +17,8 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SysUserGroupLookupRepository extends
-        JpaRepository<SysUserGroupLookup, Integer>,
-        PagingAndSortingRepository<SysUserGroupLookup, Integer>,
+        JpaRepository<SysUserGroupLookup, Long>,
+        PagingAndSortingRepository<SysUserGroupLookup, Long>,
         QuerydslPredicateExecutor<SysUserGroupLookup> {
 
 }
